requirements.txt
pandas
gradio
python-dotenv
datasets
sqlalchemy
langchain
langchain-community
langchain-openai
openai
